maildir: If we see unwanted non-empty directories in new/ or cur/, move them to ../extra-*

+++ src/lib-storage/index/maildir/maildir-sync-index.c
@@ -98,6 +98,8 @@ static int maildir_expunge(struct maildir_mailbox *mbox, const char *path,
 	}
 	if (errno == ENOENT)
 		return 0;
+	if (errno == EISDIR)
+		return maildir_lose_unexpected_dir(box->storage, path);
 
 	mail_storage_set_critical(&mbox->storage->storage,
 				  "unlink(%s) failed: %m", path);
